requirements.txt
customtkinter
face_recognition
scikit-learn
Pillow
numpy